Clean up the name
