+++ msal/src/test/java/com/microsoft/identity/client/e2e/tests/network/nativeauth/GetTokenTests.kt
@@ -37,7 +37,7 @@ import org.junit.Test
 import org.robolectric.annotation.Config
 
 @Config(shadows = [ShadowBaseController::class])
-class GetTokenTests : NativeAuthPublicClientApplicationOriginAbstractTest() {
+class GetTokenTests : NativeAuthPublicClientApplicationAbstractTest() {
 
     private lateinit var resources: List<NativeAuthTestConfig.Resource>
